requirements.txt
streamlit
transformers
torch
openai
python-dotenv
